requirements.txt
Django==1.11.6
django-cleanup==1.0.1
django-filter==1.1.0
djangorestframework==3.7.1
Markdown==2.6.9
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytesseract==0.1.7
pytz==2017.2
